Remove code checking for test prefix
Removing the code added in 2018 which checked if a test system was attempting
to mint outside the test DOI prefix.

Its possible it could be reworked by adding another option in which a custom
test prefix is entered - but someone with the need can be the one to do that
work.

Previous commit: fa34723f2f94ff37d9291612ae2ea91f832c7fed

Closes #48